Sidebar: move modal btns to sidebar folder
