+++ mypets-frontend/components/Layouts/Sidebar/LoginModalBtn/LoginModalBtn.js
@@ -1,45 +1,42 @@
-import { useContext, useState } from "react";
+import { useContext, useState, useEffect } from "react";
 import {
+  Button,
   useDisclosure,
   Modal,
   ModalOverlay,
   ModalContent,
   ModalHeader,
-  ModalFooter,
   ModalBody,
   ModalCloseButton,
+  Box,
   FormControl,
   FormLabel,
   FormErrorMessage,
-  FormHelperText,
-  VStack,
-  Box,
+  useToast,
 } from "@chakra-ui/react";
 import { Formik, Form, Field } from "formik";
 const Yup = require("yup");
 require("yup-password")(Yup);
 
-import AuthContext from "../../context/AuthContext";
-import MypetsBtn from "../Common/MypetsBtn/MypetsBtn";
+import AuthContext from "../../../../context/AuthContext";
+import MypetsBtn from "../../../Common/MypetsBtn/MypetsBtn";
+// import EmailInputGroup from "../Form/EmailInputGroup/EmailInputGroup";
+// import PasswordInputGroup from "../PasswordInputGroup/PasswordInputGroup";
 import LoginSocialBtnGroup from "../LoginSocialBtnGroup/LoginSocialBtnGroup";
+// import ForgetPasswordModalBtn from "../Form/ForgetPasswordModalBtn/ForgetPasswordModalBtn";
 
-function SignupModalBtn() {
+function LoginModalBtn() {
+  const toast = useToast();
   const { isOpen, onOpen, onClose } = useDisclosure();
-  const { registerUser } = useContext(AuthContext);
+  const { loginUser } = useContext(AuthContext);
 
-  const handleSubmit = (values, actions) => {
-    try {
-      actions.setSubmitting(true);
-      const user = registerUser(values.email, values.password);
-    } catch (err) {
-      console.error(err);
-    }
-    onClose();
+  const handleSubmit = async (values, actions) => {
+    actions.setSubmitting(true);
+    const status = await loginUser(values, toast);
+    actions.setSubmitting(false);
   };
 
-  const nonEmptyRegex = /(.|\s)*\S(.|\s)*/;
-
-  const signupSchema = Yup.object().shape({
+  const loginSchema = Yup.object().shape({
     email: Yup.string()
       .email("Invalid email")
       .required("Your email is required"),
@@ -49,25 +46,21 @@ function SignupModalBtn() {
       .min(8, "Password must contain at least 8 characters")
       .minNumbers(1, "Password must contain at least 1 digit")
       .minSymbols(1, "Password must contain at least 1 symbol"),
-    password2: Yup.string().oneOf(
-      [Yup.ref("password"), null],
-      "Passwords must match"
-    ),
   });
 
   return (
     <>
-      <MypetsBtn btnText="Sign up" mx={0} onClick={onOpen} />
+      <MypetsBtn btnText="Log in/Sign up" mx={0} onClick={onOpen} />
       <Modal isOpen={isOpen} onClose={onClose} isCentered>
         <ModalOverlay />
         <ModalContent mx={{ base: 4 }}>
-          <ModalHeader>Sign up</ModalHeader>
+          <ModalHeader>Log in/Sign up</ModalHeader>
           <ModalCloseButton />
           <ModalBody mb={2}>
             {/* <Formik
-                            initialValues={{ email: '', password: '', password2: '' }}
+                            initialValues={{ email: '', password: ''}}
                             onSubmit={handleSubmit}
-                            validationSchema={signupSchema}
+                            validationSchema={loginSchema}
                         >
                             {(props) => (
                                 <Form>
@@ -82,7 +75,7 @@ function SignupModalBtn() {
                                             )}
                                         </Field>
                                     </Box>
-                                    <Box mb={4}>
+                                    <Box mb={2}>
                                         <Field name='password' >
                                             {({ field, form }) => (
                                                 <FormControl isInvalid={form.errors.password && form.touched.password}>
@@ -93,22 +86,15 @@ function SignupModalBtn() {
                                             )}
                                         </Field>
                                     </Box>
-                                    <Box mb={4}>
-                                        <Field name='password2' >
-                                            {({ field, form }) => (
-                                            <FormControl isInvalid={form.errors.password2 && form.touched.password2}>
-                                                <FormLabel>Confirm password</FormLabel>
-                                                <PasswordInputGroup 
-                                                    field={field} 
-                                                    valid={!form.errors.password2 && form.touched.password2} 
-                                                    id='password2' 
-                                                />
-                                                <FormErrorMessage>{form.errors.password2}</FormErrorMessage>
-                                            </FormControl>
-                                            )}
-                                        </Field>
-                                    </Box>
-                                    <MypetsBtn w='100%' btnText='Sign up' mx={0} mt={8} isLoading={props.isSubmitting} type='submit'/>
+                                    <ForgetPasswordModalBtn />
+                                    <MypetsBtn
+                                        btnText='Log in'
+                                        w='100%'
+                                        mx={0}
+                                        mt={4}
+                                        isLoading={props.isSubmitting}
+                                        type='submit'
+                                    />
                                 </Form>
                             )}
                         </Formik> */}
@@ -120,4 +106,4 @@ function SignupModalBtn() {
   );
 }
 
-export default SignupModalBtn;
+export default LoginModalBtn;
